Implementando el to fixed

+++ backend/Modulos/funcion_nativa.py
@@ -26,8 +26,8 @@ class FuncionNativa():
         return array1 + array2
 
     # Hace un round al numero especificado con el numero de dcimales especificado
-    def hacer_to_fixed(self, numero: float, no_decimales: int):
-        return round(numero, no_decimales)
+    def hacer_to_fixed(self, numero: float, no_decimales: float):
+        return round(numero,  int(no_decimales))
 
     def hacer_to_exponential(self, numero: float, numeroExponencial: int):
         xx = f"{numero:.{numeroExponencial}E} "
